Refactoring: improved code quality.

+++ src/org/parabuild/ci/webui/BuildRunChangesDiffPage.java
@@ -82,8 +82,8 @@ public final class BuildRunChangesDiffPage extends AbstractBuildRunResultPage im
 
     // run search if params are OK
     if (startBuildNumber != null && endBuildNumber != null
-            && startBuildNumber.intValue() > 0 && endBuildNumber.intValue() > 0) {
-      pnlDiff.display(startBuildNumber.intValue(), endBuildNumber.intValue());
+            && startBuildNumber > 0 && endBuildNumber > 0) {
+      pnlDiff.display(startBuildNumber, endBuildNumber);
     }
 
     return makeDoneResult(buildRun);
